Added option to use default kicad directory for schematic output.
When running the eeschema export command on my computer kicad would create
the output directory normally however and extra folder with an invalid folder
name would also be created in that directory. The pdf schematic would be put
into that invalid folder making it inaccessable from the command line. The
easiest way I found to get around this was to just leave the output directory
blank so that Kicad would put the pdf in the project directory instead.